feat(opl): widen OPL pipeline to int32_t for clean gain and precision
Replace the int16_t intermediate in the OPL output path with int32_t
throughout: emu8950 mix_output_raw, OPL_calc_buffer, OPL_Pico_simple,
and get_opl_sample.

Remove the >> 1 attenuation in _MO() that was required to prevent
int16_t overflow on multi-channel sums. This eliminates 1 bit of
permanent precision loss per channel sample.

Use int64_t intermediate in scale_sample() to prevent int32_t overflow
when opl_volume is boosted above 1.0x (was undefined behaviour at high
volumes with the old approach).

Apply a 6x OPL gain constant in CMD_OPLVOL to compensate for the
removed attenuation and provide comfortable headroom. Per-channel peak
rises from ~25% to ~75% of full scale at 100% oplvol; the recommended
operating point is ~87% oplvol, leaving ~13% of range for boosting
quiet music. The volume setting now correctly attenuates the full-
precision signal rather than scaling an already-clipped intermediate.

+++ sw/audio/volctrl.h
@@ -37,7 +37,7 @@ extern int32_t psg_volume;
 extern int32_t set_volume_scale (uint8_t percent);
 
 static inline int32_t scale_sample (int32_t sample, const int32_t scale, const bool clamp) {
-    sample = (sample * scale) >> 16;
+    sample = (int32_t)(((int64_t)sample * scale) >> 16);
 
     if (clamp)
         sample = clamp16(sample);
